b2sdk.v1.sync refactored to reflect b2sdk.sync structure
